Split the documentation out of the top-level README (#331)
Moves most of the documentation into a dedicated `docs/README` to make
it easier to reach important content in the top-level readme like
"getting involved".

There are some minor changes to the heading levels here and link
changes, but otherwise the content is the same. Don't go deep on the
getting started section, that's changing in #329