fix link to CLI docs (#341)
## Summary

The file is `cli.md`, not `commands.md`.

Not sure if there was an incomplete intention to rename the file here,
but for the moment I'm just fixing the link; renaming the file will
require adjusting at least the release script and possibly changing it
in the ruff repo as well.

## Test Plan

On this branch, the link works.