Contributor guide: document sources of autogenerated reference docs (#490)
## Summary

Explain in `CONTRIBUTING.md` that some ty help text and reference
documentation result from source code that lives in Ruff, and point to
sources, to help tech writer contributors.

Inspired by my stumble in #478 and subsequent pointers in
https://github.com/astral-sh/ruff/pull/18246 .

## Test Plan

I used `git grep` to search the Ruff repository for several chunks of
the prose in `docs/reference/` to verify which source code contains text
that turns into the autogenerated docs. I pushed my branch to GitHub and
manually checked that the formatting and hyperlinks were correct.