Only add the actual schema in schemastore PRs (#1391)
Last time I ran this script, due to what I assume was a `npm` version
mismatch, the `package-lock.json` file was updated while running `npm
install` in the `schemastore`. Due to the use of `git commit -a`, it was
accidentally included in the commit for the semi-automated schemastore
PR. The solution here is to only add the actual file that we want to
commit.

I'll open similar PRs for the respective scripts in ruff and uv once
this is accepted.