Add a "type checking" page (#1915)
I was a bit unhappy that the "Introduction" had to explain concepts like
file selection and environment discovery. I also think the "Concepts"
sections didn't really make sense as "Installation" -> "Configuration",
I wanted a page for _using_ the type checker in between. This adds a new
page for "Type checking" and moves some of the content out of the
"Introduction". The "Features / Type checking" page is renamed to
"Features / Type system" to disambiguate — which I think is more
appropriate for the target readers anyway.

I think in the long run, we'll want to add "Guides" and have a "Type
checking your project", "Type checking a script", etc., but I think
that's out of scope for the beta.